Split return target list into own file
