+++ Runtime/SGBFontTarget.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2edca799e17e2247966f7719917a23c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
